fix(share): file architecture
- Create SharedableDetailView
- Change icon of share
